fixed send near zero floating point number throught http
There was an error when the "buildHttpPayload" method invoked "_FloatToChar"
This was due to both methods having an incorrect buffer size for storing the
pretended data.

Explanation:
_FloatToChar method had a 20 bytes long buffer for storing a double data type number with 16
precision digits, consider the number 100.23/200000000 = 5.0115000000000006e-07 being used as input
for this function, then in the 20th position of this number represented as a string, there would be a
number "0"(the one preceeding the 7 on the exponential part)  instead of the termination character '\0'

Fix: increase buffer size from 20 to 25 in both of the method mentioned before, this fix consider
at least 2 digits long exponential part, i.e 5.0115000000000006e-99 and of course the termination
character.